newTree

+++ ppm/src/package1/Example.scala
@@ -132,10 +132,10 @@ object Manipulation{
       case QEmpty => QEmpty
       case QLeaf((value,color: Color)) => QLeaf(c2,color)
       case QNode(value,one,two,three,four) => {
-        val cOne = ((c2._1._1,c2._1._2),((((c2._1._1+c2._2._1)/2) -0.5).toInt,(((c2._1._2+c2._2._2)/2) -0.5).toInt))
+        val cOne = ((c2._1._1,c2._1._2),((((c2._1._1+c2._2._1)/2).toDouble -0.5).toInt,(((c2._1._2+c2._2._2)/2).toDouble -0.5).toInt))
         val cTwo = ((cOne._2._1+1,c2._1._2),(c2._2._1,cOne._2._2))
         val cThree = ((c2._1._1,cOne._2._2 + 1),(cOne._2._1,c2._2._2))
-        val cFour = (((((c2._1._1+c2._2._1)/2) +0.5).toInt,(((c2._1._2+c2._2._2)/2) +0.5).toInt),(c2._2._1,c2._2._2))
+        val cFour = (((((c2._1._1+c2._2._1)/2).toDouble +0.5).toInt,(((c2._1._2+c2._2._2)/2).toDouble +0.5).toInt),(c2._2._1,c2._2._2))
         QNode(c2,newQTree(one,cOne),newQTree(two,cTwo),newQTree(three,cThree),newQTree(four,cFour))
         //QNode(cords(qt2),newQTree(one,QLeaf(sumCoords(cords(qt2),cords(one)),c)),newQTree(two,QLeaf(sumCoords(cords(qt2),cords(two)),c)),newQTree(three,QLeaf(sumCoords(cords(qt2),cords(three)),c)),newQTree(four,QLeaf(sumCoords(cords(qt2),cords(four)),c)))
       }
